+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.2.0
 
+New Unit Generators
+
+* pink.oscillators
+  * pulse - pulse generator
 
 ## 0.1.0